fix `test_extract_volume_patches_same_padding` casting

+++ keras/src/ops/image_test.py
@@ -2583,7 +2583,7 @@ class ExtractVolumePatchesTest(testing.TestCase):
     @parameterized.named_parameters(named_product(dtype=FLOAT_DTYPES))
     def test_extract_volume_patches_same_padding(self, dtype):
         volume = np.random.rand(1, 33, 33, 33, 1)
-        volume = volume.astype(np.float32)
+        volume = volume.astype(dtype)
         patches = kimage.extract_volume_patches(
             volume, size=(4, 4, 4), strides=(4, 4, 4), padding="same"
         )
